manual buffer
